chore: sanitize identifier before creating contact from backend
